+++ packages/cli-plugin-router/template/template-react/src/router/index.jsx
@@ -1,6 +1,12 @@
 import React, { lazy } from 'react';
 import { Navigate, RouteObject } from 'react-router-dom';
+import Home from '../pages/Home';
 
-const routes = [];
+const routes = [
+  {
+    path: '/',
+    element: <Home />
+  }
+];
 
 export default routes;