+++ packages/add/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @micro-cli/add
 
+## 1.0.7
+
+### Patch Changes
+
+- fix template bug
+- Updated dependencies
+  - @micro-cli/shared-utils@1.0.7
+
 ## 1.0.5
 
 ### Patch Changes