+++ packages/cli-plugin-router/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @micro-cli/cli-plugin-router
 
+## 1.0.7
+
+### Patch Changes
+
+- fix template bug
+- Updated dependencies
+  - @micro-cli/create@1.0.7
+  - @micro-cli/shared-utils@1.0.7
+
 ## 1.0.6
 
 ### Patch Changes