+++ packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @micro-cli/core
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @micro-cli/add@1.0.7
+  - @micro-cli/create@1.0.7
+  - @micro-cli/shared-utils@1.0.7
+
 ## 1.0.6
 
 ### Patch Changes