+++ packages/shared-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @micro-cli/shared-utils
 
+## 1.0.7
+
+### Patch Changes
+
+- fix template bug
+
 ## 1.0.5
 
 ### Patch Changes